+++ src/server/settings/settings-procedures.ts
@@ -2,7 +2,7 @@ import { updatePassword } from "../auth/auth"
 import { resumeQueuedEpisodeGeneration } from "../feeds/feed-podcast"
 import { restartServer } from "../serve"
 import { procedure } from "../trpc/trpc"
-import { checkForUpdateNow } from "../updater"
+import { checkForUpdateNow } from "../update-check"
 import { getEpisodeGenerationSettings, getServerBaseUrl, getServerSettings, listImageDescriptionSettings, listProviderSettings, saveEpisodeGenerationSettings, saveImageDescriptionSettings, saveProviderSettings, saveServerSettings } from "./settings-repository"
 import { SettingsInput } from "./settings-types"